# Some performance optimizations