--- a/go.mod
+++ b/go.mod
@@ -1,3 +1,5 @@
 module github.com/gohugoio/hashstructure
 
 go 1.18
+
+require github.com/cespare/xxhash/v2 v2.3.0